refactor: Rename HilingualSegmentedToggle to HilingualSegmentedButton and update usage in DiaryViewModeToggle
